Fix sensors saturation by converting to int32_t

+++ src/main/sensors/boardalignment.h
@@ -23,5 +23,5 @@ typedef struct boardAlignment_s {
     int16_t yawDegrees;
 } boardAlignment_t;
 
-void alignSensors(int16_t *src, int16_t *dest, uint8_t rotation);
+void alignSensors(int32_t *src, int32_t *dest, uint8_t rotation);
 void initBoardAlignment(boardAlignment_t *boardAlignment);
